pyproject.toml
[tool.poetry]
name = "e-commerce api"
version = "0.1.0"
description = "This is a simple E-commerce app using Django Rest Framework"
authors = ["cla-bit <doublep098@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = ["django rest framework", "django", "python", "api", "ecommerce"]
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Development Status :: 5 - Production/Stable",
    "Natural Language :: English",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[license]
text = "Copyright ©2024 Peter Mbachu"

[tool.poetry.urls]
"Source Code" = "https://github.com/cla-bit/Backend-Test-.git"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
pylint = "^3.2.5"
flake8 = "^7.1.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--doctest-modules"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
